requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
hyperframe==6.0.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.28
langchain-mistralai==0.2.4
langsmith==0.2.4
multidict==6.1.0
mypy-extensions==1.0.0
orjson==3.10.12
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
postgrest==0.18.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.0.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
yarl==1.18.3